backend/requirements.txt
fastapi
uvicorn[standard]
sqlalchemy
psycopg2-binary
pydantic
python-dotenv
openai
google-generativeai
pandas
python-multipart
celery
redis
langdetect
